backend/requirements.txt
django==5.0.3
celery==5.3.6
django-celery-beat==2.6.0
gunicorn==21.2.0
kombu==5.3.5
psycopg2-binary==2.9.8
redis==5.0.2
djangorestframework==3.14.0
django-allauth==0.52.0
dj-rest-auth[with_social]==4.0.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

black==24.2.0 ; extra == 'dev'
mypy==1.9.0 ; extra == 'dev'
pytest-cov==4.1.0 ; extra == 'dev'
pytest==8.1.1
pytest-django==4.8.0